requirements.txt
scapy
yfinance
pylama
black
pylint
netmiko
python-dotenv